PENDING: binman: add key-path to etypes
The image binaries generated by build are signed using a key file.
Currently by default the example key stored in
arch/arm/mach-k3/keys/custMpk.pem is used for this. Changing the key for
any purpose (testing, production environment, etc.) requires swapping
the key here.

This patch adds a key-path etype to binman along with a Makefile
variable KEYFILE which can be used to provide the path to a key file
directly with build commands.

Signed-off-by: T Pratham <t-pratham@ti.com>